+++ TemplateRESTful.Persistence/Seeding/Users/DefaultBasicUser.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using TemplateRESTful.Domain.Enums.Account;
-using TemplateRESTful.Domain.Models.Account;
+using TemplateRESTful.Domain.Models.Entities;
 
 namespace TemplateRESTful.Persistence.Seeding
 {