+++ TemplateRESTful.Web/Areas/Admin/Controllers/RolesController.cs
@@ -1,8 +1,9 @@
-﻿using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
-using TemplateRESTful.Domain.Models.Account;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+
+using TemplateRESTful.Domain.Models.Entities;
 
 namespace TemplateRESTful.Web.Areas.Admin.Controllers
 {