+++ TemplateRESTful.Web/Areas/Identity/IdentityHostingStartup.cs
@@ -1,11 +1,7 @@
 ﻿using System;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.UI;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using TemplateRESTful.Domain.Models.Account;
+
+using TemplateRESTful.Domain.Models.Entities;
 using TemplateRESTful.Persistence.Storage;
 
 [assembly: HostingStartup(typeof(TemplateRESTful.Web.Areas.Identity.IdentityHostingStartup))]