+++ TemplateRESTful.Web/Areas/Identity/Pages/Account/ExternalLogin.cshtml.cs
@@ -6,16 +6,16 @@ using System.Security.Claims;
 using System.Text;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
+
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.UI.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.Logging;
-using TemplateRESTful.Domain.Models.Account;
-using TemplateRESTful.Web.Implementation;
+
+using TemplateRESTful.Domain.Models.Entities;
 
 namespace TemplateRESTful.Web.Areas.Identity.Pages.Account
 {