+++ TemplateRESTful.Web/Areas/Identity/Pages/Account/Logout.cshtml.cs
@@ -1,13 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Logging;
-using TemplateRESTful.Domain.Models.Account;
+
+using TemplateRESTful.Domain.Models.Entities;
 using TemplateRESTful.Web.Implementation;
 
 namespace TemplateRESTful.Web.Areas.Identity.Pages.Account