+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 Version 1.0 (13 December 2023)
 
 - Initial release
+
+Version 1.2.1 (21 December 2023)
+
+- Tweak: Set Json Web Token for a logged in User
+- Tweak: Rename Controller to be used by different users
+- Tweak: Use nLog package to trace and debug start process
+- Tweak: Refactor service that tracks User login activity
+- Fix: Error after calling mediator service on login page